requirements.txt
flask>=1.0.0
google-cloud-firestore>=1.4.0
